requirements.txt
aiofiles==0.6.0
attrs==21.2.0
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
fastapi==0.62.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.1
python-dotenv==0.15.0
python-multipart==0.0.5
requests==2.25.0
six==1.16.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.12.3
